task3: now casts equality as float to enable summation

+++ supervised_learning/tensorflow/3-calculate_accuracy.py
@@ -18,5 +18,7 @@ def calculate_accuracy(y, y_pred):
         number of correct predcitions / total predictions
     """
 
-    return tf.divide(tf.math.reduce_sum(tf.math.equal(y, y_pred)),
-                     len(y))
+    return tf.divide(tf.math.reduce_sum(
+            tf.cast(tf.math.equal(y, y_pred),
+                    tf.float32),
+            len(y)))
